Fixed bug in RegulatorGeneAssociationEnrichmentResult.h:
Scores are now saved as double.

+++ libraries/genetrail2/regulation/RegulatorGeneAssociationEnrichmentResult.h
@@ -96,7 +96,7 @@ struct RegulatorEnrichmentResult
 		writer.String(name.c_str());
 
 		writer.String("score");
-		writer.Int(score);
+		writer.Double(score);
 
 		writer.String("hits");
 		writer.Double(hits);
